pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boltiotai>=0.0.3",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.21.1",
    "gunicorn>=23.0.0",
    "openai>=1.91.0",
    "psycopg2-binary>=2.9.10",
]
